the task 1-6

+++ 0x00-hello_world/6-size.c
@@ -7,9 +7,9 @@ main ()
 	char d;
 	float e;
 	printf("Size of an int is: %lu byte(s)\n", (unsigned long)sizeof(a));
-	printf("Size of a long int is: %lu byte(s)\n",(unsigned long)sizeof(b));
-	printf("Size of a long long int is: %lu byte(s)\n",(unsigned long)sizeof(c));
-	printf("Size of a char is: %lu byte(s)\n",(unsigned long)sizeof(d));
-	printf("Size of a float is: %lu byte(s)\n",(unsigned)sizeof(e));
-	return(0);
+	printf("Size of a long int is: %lu byte(s)\n", (unsigned long)sizeof(b));
+	printf("Size of a long long int is: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of a char is: %lu byte(s)\n", (unsigned long)sizeof(d));
+	printf("Size of a float is: %lu byte(s)\n", (unsigned long)sizeof(e));
+	return (0);
 }
